visualize_test: Split into external and non-external tests
