+++ testdata/dig_as_two.dot
@@ -3,7 +3,7 @@ digraph {
 	graph [compound=true];
 	
 		subgraph cluster_0 {
-			label = "go.uber.org/dig";
+			label = "go.uber.org/dig_test";
 			constructor_0 [shape=plaintext label="TestVisualize.func4.1"];
 			
 			"io.Reader" [label=<io.Reader>];
