+++ testdata/error.dot
@@ -1,51 +1,51 @@
 digraph {
 	rankdir=RL;
 	graph [compound=true];
-	"[type=dig.t1 group=g1]" [shape=diamond label=<dig.t1<BR /><FONT POINT-SIZE="10">Group: g1</FONT>> color=red];
-		"[type=dig.t1 group=g1]" -> "dig.t1[group=g1]0";
+	"[type=dig_test.t1 group=g1]" [shape=diamond label=<dig_test.t1<BR /><FONT POINT-SIZE="10">Group: g1</FONT>> color=red];
+		"[type=dig_test.t1 group=g1]" -> "dig_test.t1[group=g1]0";
 		
-	"[type=dig.t2 group=g2]" [shape=diamond label=<dig.t2<BR /><FONT POINT-SIZE="10">Group: g2</FONT>> color=orange];
-		"[type=dig.t2 group=g2]" -> "dig.t2[group=g2]0";
-		"[type=dig.t2 group=g2]" -> "dig.t2[group=g2]2";
+	"[type=dig_test.t2 group=g2]" [shape=diamond label=<dig_test.t2<BR /><FONT POINT-SIZE="10">Group: g2</FONT>> color=orange];
+		"[type=dig_test.t2 group=g2]" -> "dig_test.t2[group=g2]0";
+		"[type=dig_test.t2 group=g2]" -> "dig_test.t2[group=g2]2";
 		
 	
 		subgraph cluster_0 {
-			label = "go.uber.org/dig";
+			label = "go.uber.org/dig_test";
 			constructor_0 [shape=plaintext label="TestVisualize.func7.1"];
 			color=orange;
-			"dig.t3[name=n3]" [label=<dig.t3<BR /><FONT POINT-SIZE="10">Name: n3</FONT>>];
-			"dig.t2[group=g2]0" [label=<dig.t2<BR /><FONT POINT-SIZE="10">Group: g2</FONT>>];
+			"dig_test.t3[name=n3]" [label=<dig_test.t3<BR /><FONT POINT-SIZE="10">Name: n3</FONT>>];
+			"dig_test.t2[group=g2]0" [label=<dig_test.t2<BR /><FONT POINT-SIZE="10">Group: g2</FONT>>];
 			
 		}
 		
 		
-			constructor_0 -> "[type=dig.t1 group=g1]" [ltail=cluster_0];
+			constructor_0 -> "[type=dig_test.t1 group=g1]" [ltail=cluster_0];
 		
 		subgraph cluster_1 {
-			label = "go.uber.org/dig";
+			label = "go.uber.org/dig_test";
 			constructor_1 [shape=plaintext label="TestVisualize.func7.2"];
 			color=orange;
-			"dig.t4" [label=<dig.t4>];
+			"dig_test.t4" [label=<dig_test.t4>];
 			
 		}
 		
-			constructor_1 -> "dig.t3[name=n3]" [ltail=cluster_1];
+			constructor_1 -> "dig_test.t3[name=n3]" [ltail=cluster_1];
 		
 		
-			constructor_1 -> "[type=dig.t2 group=g2]" [ltail=cluster_1];
+			constructor_1 -> "[type=dig_test.t2 group=g2]" [ltail=cluster_1];
 		
 		subgraph cluster_2 {
-			label = "go.uber.org/dig";
+			label = "go.uber.org/dig_test";
 			constructor_2 [shape=plaintext label="TestVisualize.func7.4"];
 			color=red;
-			"dig.t1[group=g1]0" [label=<dig.t1<BR /><FONT POINT-SIZE="10">Group: g1</FONT>>];
-			"dig.t2[group=g2]2" [label=<dig.t2<BR /><FONT POINT-SIZE="10">Group: g2</FONT>>];
+			"dig_test.t1[group=g1]0" [label=<dig_test.t1<BR /><FONT POINT-SIZE="10">Group: g1</FONT>>];
+			"dig_test.t2[group=g2]2" [label=<dig_test.t2<BR /><FONT POINT-SIZE="10">Group: g2</FONT>>];
 			
 		}
 		
 		
-	"dig.t2[group=g2]0" [color=orange];
-	"dig.t4" [color=orange];
-	"dig.t1[group=g1]0" [color=red];
+	"dig_test.t2[group=g2]0" [color=orange];
+	"dig_test.t4" [color=orange];
+	"dig_test.t1[group=g1]0" [color=red];
 	
 }
