+++ testdata/missingDep.dot
@@ -2,6 +2,6 @@ digraph {
 	rankdir=RL;
 	graph [compound=true];
 	
-	"dig.t1" [color=red];
+	"dig_test.t1" [color=red];
 	
 }
