+++ testdata/named.dot
@@ -3,22 +3,22 @@ digraph {
 	graph [compound=true];
 	
 		subgraph cluster_0 {
-			label = "go.uber.org/dig";
+			label = "go.uber.org/dig_test";
 			constructor_0 [shape=plaintext label="TestVisualize.func3.1"];
 			
-			"dig.t1[name=bar]" [label=<dig.t1<BR /><FONT POINT-SIZE="10">Name: bar</FONT>>];
-			"dig.t2[name=baz]" [label=<dig.t2<BR /><FONT POINT-SIZE="10">Name: baz</FONT>>];
+			"dig_test.t1[name=bar]" [label=<dig_test.t1<BR /><FONT POINT-SIZE="10">Name: bar</FONT>>];
+			"dig_test.t2[name=baz]" [label=<dig_test.t2<BR /><FONT POINT-SIZE="10">Name: baz</FONT>>];
 			
 		}
 		
-			constructor_0 -> "dig.t3[name=foo]" [ltail=cluster_0];
+			constructor_0 -> "dig_test.t3[name=foo]" [ltail=cluster_0];
 		
 		
 		subgraph cluster_1 {
-			label = "go.uber.org/dig";
+			label = "go.uber.org/dig_test";
 			constructor_1 [shape=plaintext label="TestVisualize.func3.2"];
 			
-			"dig.t3[name=foo]" [label=<dig.t3<BR /><FONT POINT-SIZE="10">Name: foo</FONT>>];
+			"dig_test.t3[name=foo]" [label=<dig_test.t3<BR /><FONT POINT-SIZE="10">Name: foo</FONT>>];
 			
 		}
